Fix flake8 F codes
