+++ astroquery/svo_fps/core.py
@@ -1,4 +1,3 @@
-import numpy as np
 import requests
 import io
